--- db/models/newsAndEvents.js
@@ -1,41 +1,0 @@
-/**
- * This file defines the Sequelize schema for News and Events.
- *
- * @summary   defines schema for News and Events.
- */
-const Sequelize = require("sequelize");
-const db = require("../configDB");
-
-module.exports = db.define(
-    "newsAndEvents",
-    {
-        // By default, there will be a column called 'id' that will
-        // auto-increment an integer value
-        title: {
-            type: Sequelize.STRING(20),
-            allowNull: false,
-        },
-        description: {
-            type: Sequelize.STRING(175),
-            allowNull: true,
-            defaultValue: "",
-        },
-        imageLink: {
-            type: Sequelize.STRING(500),
-            allowNull: false,
-        },
-        redirectLink: {
-            type: Sequelize.STRING(500),
-            allowNull: false,
-        },
-        openInSameTab: {
-            type: Sequelize.BOOLEAN,
-            allowNull: true,
-            defaultVale: false,
-        },
-    },
-    {
-        // createdAt & updatedAt columns will be added/self-mantained by table
-        timestamps: true,
-    }
-);
