+++ seeds/initializeTables.js
@@ -8,7 +8,7 @@ const initBranchesAndChapters = require("./models/branchesAndChapters");
 const initAdminAccounts = require("./models/adminAccounts");
 const initVolunteerInterests = require("./models/volunteerInterests");
 const initEmailList = require("./models/emaillist");
-const initNewsEvents = require("./models/newsAndEvents");
+const initExecCommittee = require("./models/execCommittees");
 
 module.exports = () => {
     Promise.all([
@@ -22,7 +22,7 @@ module.exports = () => {
         initMemberships(),
         initVolunteerInterests(),
         initEmailList(),
-        initNewsEvents(),
+        initExecCommittee(),
     ])
         .then(() => {
             console.log("Done!");
