+++ server.js
@@ -53,7 +53,6 @@ app.use("/newsletters", require("./routes/newsletters"));
 app.use("/adminAccounts", require("./routes/adminAccounts"));
 app.use("/contact", require("./routes/contact"));
 app.use("/branchesAndChapters", require("./routes/branchesAndChapters"));
-app.use("/newsAndEvents", require("./routes/newsAndEvents"));
 app.use("/execCommittees", require("./routes/execCommittees"));
 
 app.get("/", (req, res) => {
